move locale configuration into user package
